Add material property to product info section
A new property, namely "material", has been added to the product info section. This involved updates on the Product InfoSection view, the product details page, the product service and the product type definitions. The material property is now displayed on the front end and is also fetched from the backend.